Add active tasks observability.
Follow-up on the new strategy `limit-active-tasks`,
allow operators to follow the evolution of the
number of active tasks in the system through fly
and metrics.

Show active tasks counter when `fly workers` is invoked
with the --details option.
Expand fly tests to cover for the above feature.

Add active tasks to the workers metrics and the
Prometheus exporter.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>