restore string coercion behavior for task env
this was previously done by the SanitizeDecodeHook which has been
eliminated.

this is a little bit legacy, and I'm not sure we should support it
forever.

on the plus side, we now have a proper type for task env, which feels a
little cleaner than just throwing maps around, and we have a explicit
type for coerced strings which is a bit nicer to implement than via
reflection with the decode hook.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>