atc: algorithm refactorings and remove skipped
The skipped column in the next build inputs was removed because we will
now only save either a version or an error in that table. So the table
will either contain a fully resolved set of versions for all its inputs
or a few errors for inputs and all the other inputs won't have a row
populated. We used to fill out the table with partially resolved inputs
when there were errors (with versions that we were able to partially
able to resolve to and also skipped inputs that were never reached
within the same resolve call). This was removed to only populate errors
when the resolve failed because we don't expect to show the partially
resolved versions to the users, so adding this feature is useless at
this time. It also added complexity to the algorithm.

[#3602]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>