Do not trigger Garden error when asking for container property.
When querying Garden for a property that does not exist, Garden
logs an error message. This error message pollutes the Worker
logs as it doesn't really indicate that something wrong has
been going on.

By querying all properties and checking if the interested one
is among them we obtain the same result while avoiding Garden
to log an error.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>