feat[gc/build-log-collector]: Added minimum_succeeded_builds to build_log_retention.
When minimum_succeeded_builds is specified, at least such success builds' logs will
be retained, which will be helpful for developing and debugging a job.

This commit changes the original algorithm of build log collector a little bit:
- no longer specially handle build id 1
- use ReapTime field to decide if a build's log has been reaped
- FirstLoggedBuildID now points to the first build whoes log is not reapped

Unit tests are updated.

If some user requests to support minimum_failed_builds in future, it would be easy
to support.

Resove #4026
Close #4188

Signed-off-by: Chao Li <chaol@vmware.com>