atc: refactor and cleanup algorithm
Split off the algorithm into different files representing different
components or objects within the algorithm. The flow is currently that
it starts with the algorithm compute, then it grabs new resolvers from
the resolver file and uses those resolvers (which might be the pinned,
individual or group resolvers) to resolve all the inputs. After it
finishes resolving, it uses the input mapper to map the candidate
versions into an input mapping object.

Also did minor cleanups everywhere to get tests to pass, fixed the flake
within the first occurrence tests and moved the gc Collector interface
to the atccmd package. This is because we no longer have an aggregate
object so we can move the interface to where it is actually used.

[#3602]

Signed-off-by: Clara Fu <cfu@pivotal.io>