Do not increase active_tasks if the container is already present.
This fixes a bug where the active_tasks counter (used in the limit-active-tasks strategy)
would be increased even if the container running the task is already present in the worker.

The above bug can manifest, for instance, if the ATC restarts.

To detect if a container is already present on a worker a small helper function is added
to Pool and queried it only when usign the limit-active-tasks strategy.

A test case is added to ensure the active_tasks counter is not increased when the container
is already present on the worker.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>