atc: set first_occ to default false
This will remove all the first_occurrence yellow arrows for all past
builds. For most people, it will be ok to remove these indicators
because they don't care about the previous builds. For the people that
do care, we are thinking of adding a separate batch command you can run
to fill in that data. This will be an expensive operation so we thought
it would be best to have it be optional (since it is purely used for
cosmetic purposes).

[#3602]

Signed-off-by: Clara Fu <cfu@pivotal.io>