atc: add inputs ready to builds
Before, the scheduled state on the builds object would be used to
determine whether a build is ready to be used within the algorithm (it
has all its build inputs moved from next build inputs to the build
inputs table). But after our change with fixing the bug with max in
flight, the scheduled boolean now is used to determine whether this
build has passed the max in flight test. We still need a boolean to help
us know when a build is ready to be used by the algorithm (has
inputs/build pipes in the db) so I added a new boolean called inputs
ready. There might be a better way to restructure the buildstarter that
will involve not needing to have two different boolean states, but this
way is more explicit with how far the build has gotten in terms of the
buildstarter code which is very important.

[#3602]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>