atc: first pass at adding retriggering builds
Added a new api endpoint for retriggering builds. The endpoint will just
save a new build in the database and give it a "retrigger_of" state with
the build id of the build that it's a retrigger of. The scheduler will
then be the one that actually takes that pending build and takes all the
inputs/pipes of the build that it's a retrigger of and creates a new
plan with those same inputs/pipes.

[#4268]

Signed-off-by: Clara Fu <cfu@pivotal.io>