atc: silently rerun the correct build
If a user tries to rerun a rerun build, it will silently create a rerun
of the build that the rerun is of. This is because we don't want a tree
of rerun builds where you can have a rerun of a rerun of a rerun of a
build etc.

[#4268]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>