atc: implement fallback for first occurrence
If the first occurrence field is nil in the database, that means that
the inputs are from old builds created before 6.0.0. If the field is
nil, we will determine the correct value for first occurrence by running
a query to figure out the value (similar to what we did before).

[#4352]

Signed-off-by: Clara Fu <cfu@pivotal.io>