Update AWS Go SDK  to version 1.25.18
AWS recently added a new feature to EKS to allow attaching IAM roles to Kubernetes service accounts. Changes were made to the AWS SDK to support this new authorization method - until the latest version of the SDK is used in Concourse, it is not possible to use Concourse in EKS where this feature is used. This change updates to the latest version of the SDK to support this.

Signed-off-by: George Webb <george.webb@voidsource.com>