worker: add initial no-op containerd backend
this commit adds preliminary code that'd back the implementation of a
garden backend.

aside from a backend implementation, it also includes a little bit of
code that could represent how we interact with containerd - via a small
interface that exposes just the methods that we need (so that we can
unit-test our backend implementation).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>