atc: create job pipes
The job pipes table will contain the information for which resources is
used by which jobs and also passed constraints between jobs. This will
help us figure out which jobs need to be requested to schedule when
certain actions happen. For example, when a new version is saved, all
jobs that use that resource will be requested to schedule and when a
build is manually triggered, only the job for that build will be
requested to schedule. This will allow us to target the specific jobs
which we know will be affected by actions, rather than having to
request schedule on the entire pipeline.

Saving the pipeline will populate the job pipes table. We also
refactored the SavePipeline method to make it more readible.

Also implemented requesting schedule for each of the actions that should
result in a schedule being requested for specific jobs.

[#4717]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Julia Pu <jpu@pivotal.io>