Allow topgun to run all tests suites
Currently for our newer release pipelines that have the refactored
version of bosh topgun (5.7.x), when we want to run all test suites we
don't specify `SUITE` and run `ginkgo` with the recursive flag. This
allows us to run all topgun suites.

The problem, that this commit resolves, is that the release referened in
the deployment yaml gets the wrong name. By moving the `-` out of the
deployment yaml, the naming will now be correct.

From CI we currently get this error:

```
Task 26668 | 19:56:13 | Error: Release 'concourse-' doesn't exist
```

because `concourse-((suite))` resolves to `concourse-` because `SUITE`
is empty. Now this commit appends the dash as part of the `((suite))`
variable.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>