atc: lazy load job configs
This commit makes it so that when a job is scanned out from the
database, we do not decrypt the config until it explicitly needs it.

Some initial work was done on the scheduler, DB and API to reflect this
change. Multiple fields from the config are also being removed from the
config because they do not need to be encrypted. Instead we are going to
store them in the database as individual columns. Examples of that is
moving public and max_in_flight out of the config in the jobs table.

Job pipes table was also renamed to job_inputs in order to be more
accurate since "job_pipes" are only the get steps of a job. We also
intended to add a job_outputs table but after some consideration, we
might not need it anymore because the dashboard does not require job
outputs.

The dashboard was a huge change in this commit, where we changed all the
data needed (job, finished build, next build and transition build)
information into a single query. This will cause some changes in the
response of the API (ListAllJobs) for the dashboard.

[#4862]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Joshua Winters <jwinters@pivotal.io>