atc: fail in non-specific error in db container creation
In `atc/worker/worker.go`'s garden's FindOrCreateContainer
implementation, there's an early return for the case when the error
returned in that creation is of a specific kind, but not for other
possible errors that could happen.

That meant that any error other than that specific one would lead to a
NPE as further steps would rely on a non-nil return.

These changes fix that by actually returning the err, making further
assumptions on "not being nil" not problematic anymore.

fixes #4932

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>