job label for prometheus build_duration metrics
includes the job name as a label in build duration metrics exposed to
prometheus so that it is possible to gain insights into where time is
being spent in a complex pipeline, observe trends and create alerts at
job level.

before this change the build_duration metric is not very useful to us, as it bundles
together all builds accross an entire pipeline, since a build duration from one
job is usually very distinct from a build of another job, bundling them
together makes it hard to extract useful information such as:

* Which job(s) in the pipeline are performing badly?
* Under what conditions cause job A to perform badly?

Or create alerts/indicators based on the metric such as:

* Job X should always run within Y minutes

By allowing the build_durations metric to contain the job label, these
things become possible.

Signed-off-by: chrisfarms <chris.farmiloe@digital.cabinet-office.gov.uk>