atc: migrate all types of serial groups
Prior to 6.0, when a user would set a pipeline it would only register
the serial groups into the jobs_serial_groups table and not the serial:
true or max in flight. This was fixed in 6.0 so we need to add a
migration to move all the possible missing groups into the table.

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>