atc: error checks that fail to create check step
Finish checks that error while trying to create a check step because it
can possibly be caught in a loop of trying to create a check that no
longer is valid. For example, the scenario we were getting into was that
it was trying to find the pipeline while creating the check step but the
pipeline id was not in the metadata so it would always error while
creating the check. Since we didn't error out the check, it would
continue to try to create that same check but would never be able to
find the pipeline.

This commit will finish the check with an error if anything goes wrong
while trying to create the check. This will allow us to get out of
situations where the check is possibly outdated or holds bogus
information.

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>