add tracing package
the tracing package is responsible for providing a thin wrapper on top
of opentelemetry[1].

by having this wrapper, we're able to protect ourselves from the
underlying changes that `opentelemetry` has been going through as it
goes through the process of getting ready for a final release.

the package is meant to be used in a two step fashion:

1. an exporter is configured (initializing the SDK to not use a noop
   tracer) through `tracing.ConfigureTracer`

2. spans are then created
  - tracing.StartSpan() & span.End()

so that code can be instrumented in no more than ~3 lines of code.

[1]: https://opentelemetry.io/

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>