atc: add tracepoint to engine build
adding a tracepoint right after ATC acquires the lock to track that
build, we're able to capture the duration of the entiire build as seen
by this ATC who's keeping track of it.

as the context propagates further down to other steps, we can correlate
the the entire build with its steps.

ps.: this commit DOES NOT address the case of an ATC being phased out in
the middle of the build tracking and having the work picked up by
another installation (although this should be possible if we own our
traceIDs).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>