containerd: add max uid/gid (spec)
in order to fully embrace user namespace, we need to be aware of the
possible user/group ids that we should use at the moment of creating the
OCI spec for the container.

this commit adds the ability to retrieve what the max UID and GID in the
current namespace are so that we can reflect that in the oci specs we
generate.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>