containerd: add rootfs manager
this poorly named interface takes care of setting up extra things that
should be in the rootfs at the moment when no more bind mounts can be
performed.

in its current implementation, it allows one to configure a directory
that does not exist there (`SetupCwd`).

this is important for the cases when one tries to run a process in a
container on a directory that does not exist.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>