containerd: add custom init
most of the flow that takes place in the operations that we perform on
containers look like this:

	1. create a container
	2. leave it running there for few moments
	3. execute a process there

however, as a container is a process with a certain set of kernel
features enabled on it to provide a sandboxed environment, we need to
have an init process there in the first place.

this commit created `cmd/init`, which is exactly that - an executable
that takes care of ... being there! it does nothing more than waiting
for a signal to come (having marked itself as not caring about any
children in that pid namespace).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>