web: fix a padding issue
We removed margins from the cards in a previous commit, because they are now
absolute positioned. This caused the team name headers and the bottom of the
dashboard to be spaced too close to other objects on the screen. This commit
adds that space back to those areas of the dashboard.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>