algorithm: avoid trying same candidates repeatedly
Say you have a pipeline like so:

  {r1}--[j1]--{r1}--[j2]
  {r2}-/

* [j1] is a job with two inputs, {r1} and {r2}
* [j2] is a downstream job which only cares about the {r1} input.

If {r2} generates new versions much more frequently than {r1}, there
will be potentially many more builds of [j1] for [j2] to consider during
the algorithm, many of which having the same version of {r1} (because it
doesn't change as often).

With an unhappy pipeline, where the `passed` constraint can't be
satisified quickly, this can result in a lot of wasteful churn as the
algorithm keeps attempting the same version of {r1} over and over until
it eventually gets to a build that has a different version of it.

If you increase the complexity of this pipeline by adding many more jobs
upstream of [j2], each with many builds of their own each because they
also have another resource which makes them run more often, this can
lead to the algorithm running for a very very long time.

To mitigate this, we can just make note of whichever previous set of
candidates we tried and failed. If we're about to recurse with the exact
same set of candidates, we can just short-circuit, saving on a lot of
wasteful recursion.

A regression test has been added, which currently completes in ~30
seconds. I don't really know how long it would run for before because I
didn't have the patience.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>