fly, go-concourse: Better errors with --team
We realized the get-team integration tests were never running because
all the actual test cases were stuck inside a BeforeEach! Fixed that up and got
a reasonable test harness running -- `ginkgo ./fly/integration` with the
get-team tests focused and `ginkgo ./go-concourse/concourse` for the client/team
unit tests.

We have been shifting around some code in
commands.helpers/rc.target/concourse.client/concourse.team, because invocations
like `target.Team().Team(teamName) are a bit of a WTF.

We introduced `FindTeam` methods to avoid breaking existing code, but we may
completely migrate calls to the existing `Team` functions to always use
`FindTeam` -- the jury is still out on this, because it would mean most fly
commands making (potentially unnecessary) extra requests to the
`/api/v1/teams/:team_name` endpoint, purely for the purpose of printing an error
if the request fails.

We removed the `found` return value from FindTeam from the
concourse.Client struct. We brought back the switch..case statement so
we can return a custom error message when we get a 404. All other errors
are returned as is. In the fly code we can now return the error from
go-concourse as-is and it should be a helpful error message ("team
'doesnotexist' not found").

concourse/concourse#5126

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>